Add CircularSeekBar AddActivity

+++ app/src/main/java/com/romain/cellarv1/controleur/Controle.java
@@ -61,7 +61,7 @@ public class Controle {
      * @param favorite
      * @param wish
      */
-    public void createWineBottle(Integer id, String country, String region, String wineColor, String domain, String appellation, Integer year, Integer apogee, Integer number, Integer estimate, String pictureLarge, String pictureSmall, byte[] imageLarge, byte[] imageSmall, Float rate, String favorite, String wish, Float lattitude, Float longitude, String timeStamp, Context context) {
+    public void createWineBottle(Integer id, String country, String region, String wineColor, String domain, String appellation, Integer year, Integer apogee, Integer number, Integer estimate, String pictureLarge, String pictureSmall, byte[] imageLarge, byte[] imageSmall, Integer rate, String favorite, String wish, Float lattitude, Float longitude, String timeStamp, Context context) {
         wineBottle = new WineBottle(id, country, region, wineColor, domain, appellation, year, apogee, number, estimate, pictureLarge, pictureSmall, imageLarge, imageSmall, rate, favorite, wish, lattitude, longitude, timeStamp);
         accesLocal.add(wineBottle);
     }
